wip: Media-Meta PoC
proof of concept for fetching metadata about the currently played song
on OE3. No real safeguards yet if the albumart couldn't be loaded and
the request is still way to naive. Also the styling and the logic around
what should be displayed when needs to be moved to a component itself