feat: Refactor Media // Live Meta-Data
Refactored the handling of Media to be based on an interface, and moved
the player-logic out of it. Also implementen logic to fetch Meta-data
and handle the redrawing of the UI based on an event, that gets
triggered based on when the current song should end. A bit of
fallback-logic if something wents wrong with the fetch, or the end-date
makes no sense. PoC ready.